+++ include/CMakeLists.txt
@@ -17,10 +17,136 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+option(WITH_OPAQUE_SETTINGS "Hide rdpSettings struct definition, only allow getter/setter access" OFF)
+
 configure_file(${CMAKE_CURRENT_SOURCE_DIR}/config/version.h.in ${CMAKE_CURRENT_BINARY_DIR}/freerdp/version.h)
 configure_file(${CMAKE_CURRENT_SOURCE_DIR}/config/build-config.h.in ${CMAKE_CURRENT_BINARY_DIR}/freerdp/build-config.h)
 configure_file(${CMAKE_CURRENT_SOURCE_DIR}/config/config.h.in ${CMAKE_CURRENT_BINARY_DIR}/freerdp/config.h)
 configure_file(${CMAKE_CURRENT_SOURCE_DIR}/config/buildflags.h.in ${CMAKE_CURRENT_BINARY_DIR}/freerdp/buildflags.h)
+
+file(STRINGS freerdp/settings_types_private.h SETTINGS_KEYS REGEX "ALIGN64[ \ta-zA-Z0-9]*")
+
+set (SETTINGS_KEYS_BOOL "")
+set (SETTINGS_KEYS_INT16 "")
+set (SETTINGS_KEYS_UINT16 "")
+set (SETTINGS_KEYS_INT32 "")
+set (SETTINGS_KEYS_UINT32 "")
+set (SETTINGS_KEYS_INT64 "")
+set (SETTINGS_KEYS_UINT64 "")
+set (SETTINGS_KEYS_STRING "")
+set (SETTINGS_KEYS_POINTER "")
+
+foreach(KEY ${SETTINGS_KEYS})
+    string(STRIP "${KEY}" TRIMMED_KEY)
+    string(REGEX MATCH "^ALIGN64[ \t ]+BOOL[ \t ]+" IS_BOOL "${TRIMMED_KEY}")
+    string(REGEX MATCH "^ALIGN64[ \t ]+INT16[ \t ]+" IS_INT16 "${TRIMMED_KEY}")
+    string(REGEX MATCH "^ALIGN64[ \t ]+UINT16[ \t ]+" IS_UINT16 "${TRIMMED_KEY}")
+    string(REGEX MATCH "^ALIGN64[ \t ]+INT32[ \t ]+" IS_INT32 "${TRIMMED_KEY}")
+    string(REGEX MATCH "^ALIGN64[ \t ]+UINT32[ \t ]+" IS_UINT32 "${TRIMMED_KEY}")
+    string(REGEX MATCH "^ALIGN64[ \t ]+INT64[ \t ]+" IS_INT64 "${TRIMMED_KEY}")
+    string(REGEX MATCH "^ALIGN64[ \t ]+UINT64[ \t ]+" IS_UINT64 "${TRIMMED_KEY}")
+    string(REGEX MATCH "^ALIGN64[ \t ]+(char|CHAR)[ \t ]*\\*[ \t ]+" IS_STRING "${TRIMMED_KEY}")
+
+    string(REGEX REPLACE ".+/\\*" "" index "${TRIMMED_KEY}")
+    string(REGEX REPLACE "[ \t/\\*]" "" index "${index}")
+
+    if (index MATCHES "^[0-9]+$")
+        if (IS_BOOL)
+            string(REGEX REPLACE "^ALIGN64[ \t ]+BOOL[ \t ]+" "" VALUE "${TRIMMED_KEY}")
+            string(FIND "${VALUE}" ";" SEMICOLON)
+            string(SUBSTRING "${VALUE}" 0 ${SEMICOLON} KEY_VALUE)
+            set(KEY_VALUE "FreeRDP_${KEY_VALUE} = ${index}")
+            list(APPEND SETTINGS_KEYS_BOOL ${KEY_VALUE})
+        elseif(IS_INT16)
+            string(REGEX REPLACE "^ALIGN64[ \t ]+INT16[ \t ]+" "" VALUE "${TRIMMED_KEY}")
+            string(FIND "${VALUE}" ";" SEMICOLON)
+            string(SUBSTRING "${VALUE}" 0 ${SEMICOLON} KEY_VALUE)
+            set(KEY_VALUE "FreeRDP_${KEY_VALUE} = ${index}")
+            list(APPEND SETTINGS_KEYS_INT16 ${KEY_VALUE})
+        elseif(IS_UINT16)
+            string(REGEX REPLACE "^ALIGN64[ \t ]+UINT16[ \t ]+" "" VALUE "${TRIMMED_KEY}")
+            string(FIND "${VALUE}" ";" SEMICOLON)
+            string(SUBSTRING "${VALUE}" 0 ${SEMICOLON} KEY_VALUE)
+            set(KEY_VALUE "FreeRDP_${KEY_VALUE} = ${index}")
+            list(APPEND SETTINGS_KEYS_UINT16 ${KEY_VALUE})
+        elseif(IS_INT32)
+            string(REGEX REPLACE "^ALIGN64[ \t ]+INT32[ \t ]+" "" VALUE "${TRIMMED_KEY}")
+            string(FIND "${VALUE}" ";" SEMICOLON)
+            string(SUBSTRING "${VALUE}" 0 ${SEMICOLON} KEY_VALUE)
+            set(KEY_VALUE "FreeRDP_${KEY_VALUE} = ${index}")
+            list(APPEND SETTINGS_KEYS_INT32 ${KEY_VALUE})
+        elseif(IS_UINT32)
+            string(REGEX REPLACE "^ALIGN64[ \t ]+UINT32[ \t ]+" "" VALUE "${TRIMMED_KEY}")
+            string(FIND "${VALUE}" ";" SEMICOLON)
+            string(SUBSTRING "${VALUE}" 0 ${SEMICOLON} KEY_VALUE)
+            set(KEY_VALUE "FreeRDP_${KEY_VALUE} = ${index}")
+            list(APPEND SETTINGS_KEYS_UINT32 ${KEY_VALUE})
+        elseif(IS_INT64)
+            string(REGEX REPLACE "^ALIGN64[ \t ]+INT64[ \t ]+" "" VALUE "${TRIMMED_KEY}")
+            string(FIND "${VALUE}" ";" SEMICOLON)
+            string(SUBSTRING "${VALUE}" 0 ${SEMICOLON} KEY_VALUE)
+            set(KEY_VALUE "FreeRDP_${KEY_VALUE} = ${index}")
+            list(APPEND SETTINGS_KEYS_INT64 ${KEY_VALUE})
+        elseif(IS_UINT64)
+            string(REGEX REPLACE "^ALIGN64[ \t ]+UINT64[ \t ]+" "" VALUE "${TRIMMED_KEY}")
+            string(FIND "${VALUE}" ";" SEMICOLON)
+            string(SUBSTRING "${VALUE}" 0 ${SEMICOLON} KEY_VALUE)
+            set(KEY_VALUE "FreeRDP_${KEY_VALUE} = ${index}")
+            list(APPEND SETTINGS_KEYS_UINT64 ${KEY_VALUE})
+        elseif(IS_STRING)
+            string(REGEX REPLACE "^ALIGN64[ \t ]+(char|CHAR)[ \t ]*\\*[ \t ]+" "" VALUE "${TRIMMED_KEY}")
+            string(FIND "${VALUE}" ";" SEMICOLON)
+            string(SUBSTRING "${VALUE}" 0 ${SEMICOLON} KEY_VALUE)
+            set(KEY_VALUE "FreeRDP_${KEY_VALUE} = ${index}")
+            list(APPEND SETTINGS_KEYS_STRING ${KEY_VALUE})
+        else()
+            string(REGEX REPLACE "^ALIGN64[ \t ]+[a-zA-Z0-9_\\*]+[ \t ]*\\*[ \t ]+" "" VALUE "${TRIMMED_KEY}")
+            string(FIND "${VALUE}" ";" SEMICOLON)
+            string(SUBSTRING "${VALUE}" 0 ${SEMICOLON} KEY_VALUE)
+            set(KEY_VALUE "FreeRDP_${KEY_VALUE} = ${index}")
+            list(APPEND SETTINGS_KEYS_POINTER ${KEY_VALUE})
+        endif()
+    endif()
+endforeach()
+
+if (NOT SETTINGS_KEYS_BOOL)
+    set(SETTINGS_KEYS_BOOL "FreeRDP_BOOL_UNUSED = -1")
+endif()
+if (NOT SETTINGS_KEYS_INT16)
+    set(SETTINGS_KEYS_INT16 "FreeRDP_INT16_UNUSED = -1")
+endif()
+if (NOT SETTINGS_KEYS_UINT16)
+    set(SETTINGS_KEYS_UINT16 "FreeRDP_UINT16_UNUSED = -1")
+endif()
+if (NOT SETTINGS_KEYS_INT32)
+    set(SETTINGS_KEYS_INT32 "FreeRDP_INT32_UNUSED = -1")
+endif()
+if (NOT SETTINGS_KEYS_UINT32)
+    set(SETTINGS_KEYS_UINT32 "FreeRDP_UINT32_UNUSED = -1")
+endif()
+if (NOT SETTINGS_KEYS_INT64)
+    set(SETTINGS_KEYS_INT64 "FreeRDP_INT64_UNUSED = -1")
+endif()
+if (NOT SETTINGS_KEYS_UINT64)
+    set(SETTINGS_KEYS_UINT64 "FreeRDP_UINT64_UNUSED = -1")
+endif()
+if (NOT SETTINGS_KEYS_STRING)
+    set(SETTINGS_KEYS_STRING "FreeRDP_STRING_UNUSED = -1")
+endif()
+if (NOT SETTINGS_KEYS_POINTER)
+    set(SETTINGS_KEYS_POINTER "FreeRDP_POINTER_UNUSED = -1")
+endif()
+string(REPLACE ";" ",\n\t" SETTINGS_KEYS_BOOL "${SETTINGS_KEYS_BOOL}")
+string(REPLACE ";" ",\n\t" SETTINGS_KEYS_INT16 "${SETTINGS_KEYS_INT16}")
+string(REPLACE ";" ",\n\t" SETTINGS_KEYS_UINT16 "${SETTINGS_KEYS_UINT16}")
+string(REPLACE ";" ",\n\t" SETTINGS_KEYS_INT32 "${SETTINGS_KEYS_INT32}")
+string(REPLACE ";" ",\n\t" SETTINGS_KEYS_UINT32 "${SETTINGS_KEYS_UINT32}")
+string(REPLACE ";" ",\n\t" SETTINGS_KEYS_INT64 "${SETTINGS_KEYS_INT64}")
+string(REPLACE ";" ",\n\t" SETTINGS_KEYS_UINT64 "${SETTINGS_KEYS_UINT64}")
+string(REPLACE ";" ",\n\t" SETTINGS_KEYS_STRING "${SETTINGS_KEYS_STRING}")
+string(REPLACE ";" ",\n\t" SETTINGS_KEYS_POINTER "${SETTINGS_KEYS_POINTER}")
+
+configure_file(${CMAKE_CURRENT_SOURCE_DIR}/config/settings_keys.h.in ${CMAKE_CURRENT_BINARY_DIR}/freerdp/settings_keys.h)
 
 file(GLOB_RECURSE PUBLIC_COMMON_HEADERS
 	LIST_DIRECTORIES false
