+++ include/config/settings_keys.h.in
@@ -1,0 +1,74 @@
+/**
+ * FreeRDP: A Remote Desktop Protocol Implementation
+ * Internal settings header for functions not exported
+ *
+ * Copyright 2023 Armin Novak <armin.novak@thincast.com>
+ * Copyright 2023 Thincast Technologies GmbH
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#ifndef FREERDP_SETTINGS_KEYS_H
+#define FREERDP_SETTINGS_KEYS_H
+
+#include <freerdp/api.h>
+
+typedef enum
+{
+    @SETTINGS_KEYS_BOOL@
+} FreeRDP_Settings_Keys_Bool;
+
+typedef enum
+{
+    @SETTINGS_KEYS_INT16@
+} FreeRDP_Settings_Keys_Int16;
+
+typedef enum
+{
+    @SETTINGS_KEYS_UINT16@
+} FreeRDP_Settings_Keys_UInt16;
+
+typedef enum
+{
+    @SETTINGS_KEYS_INT32@
+} FreeRDP_Settings_Keys_Int32;
+
+typedef enum
+{
+    @SETTINGS_KEYS_UINT32@
+} FreeRDP_Settings_Keys_UInt32;
+
+typedef enum
+{
+    @SETTINGS_KEYS_INT64@
+} FreeRDP_Settings_Keys_Int64;
+
+typedef enum
+{
+    @SETTINGS_KEYS_UINT64@
+} FreeRDP_Settings_Keys_UInt64;
+
+typedef enum
+{
+    @SETTINGS_KEYS_STRING@
+} FreeRDP_Settings_Keys_String;
+
+typedef enum
+{
+    @SETTINGS_KEYS_POINTER@
+} FreeRDP_Settings_Keys_Pointer;
+
+
+#define FreeRDP_Settings_StableAPI_MAX 5312
+
+#endif /* FREERDP_SETTINGS_KEYS_H */
