+++ libfreerdp/common/settings_getters.c
@@ -51,7 +51,7 @@ static BOOL update_string_(char** current, char* next, size_t next_len)
 	return !next || (*current != NULL);
 }
 
-BOOL freerdp_settings_get_bool(const rdpSettings* settings, size_t id)
+BOOL freerdp_settings_get_bool(const rdpSettings* settings, FreeRDP_Settings_Keys_Bool id)
 {
 	WINPR_ASSERT(settings);
 
@@ -626,7 +626,7 @@ BOOL freerdp_settings_get_bool(const rdpSettings* settings, size_t id)
 	}
 }
 
-BOOL freerdp_settings_set_bool(rdpSettings* settings, size_t id, BOOL val)
+BOOL freerdp_settings_set_bool(rdpSettings* settings, FreeRDP_Settings_Keys_Bool id, BOOL val)
 {
 	union
 	{
@@ -1398,7 +1398,7 @@ BOOL freerdp_settings_set_bool(rdpSettings* settings, size_t id, BOOL val)
 	return TRUE;
 }
 
-UINT16 freerdp_settings_get_uint16(const rdpSettings* settings, size_t id)
+UINT16 freerdp_settings_get_uint16(const rdpSettings* settings, FreeRDP_Settings_Keys_UInt16 id)
 {
 	WINPR_ASSERT(settings);
 
@@ -1451,7 +1451,7 @@ UINT16 freerdp_settings_get_uint16(const rdpSettings* settings, size_t id)
 	}
 }
 
-BOOL freerdp_settings_set_uint16(rdpSettings* settings, size_t id, UINT16 val)
+BOOL freerdp_settings_set_uint16(rdpSettings* settings, FreeRDP_Settings_Keys_UInt16 id, UINT16 val)
 {
 	union
 	{
@@ -1527,7 +1527,7 @@ BOOL freerdp_settings_set_uint16(rdpSettings* settings, size_t id, UINT16 val)
 	return TRUE;
 }
 
-INT16 freerdp_settings_get_int16(const rdpSettings* settings, size_t id)
+INT16 freerdp_settings_get_int16(const rdpSettings* settings, FreeRDP_Settings_Keys_Int16 id)
 {
 	WINPR_ASSERT(settings);
 
@@ -1541,7 +1541,7 @@ INT16 freerdp_settings_get_int16(const rdpSettings* settings, size_t id)
 	}
 }
 
-BOOL freerdp_settings_set_int16(rdpSettings* settings, size_t id, INT16 val)
+BOOL freerdp_settings_set_int16(rdpSettings* settings, FreeRDP_Settings_Keys_Int16 id, INT16 val)
 {
 	union
 	{
@@ -1565,7 +1565,7 @@ BOOL freerdp_settings_set_int16(rdpSettings* settings, size_t id, INT16 val)
 	return TRUE;
 }
 
-UINT32 freerdp_settings_get_uint32(const rdpSettings* settings, size_t id)
+UINT32 freerdp_settings_get_uint32(const rdpSettings* settings, FreeRDP_Settings_Keys_UInt32 id)
 {
 	WINPR_ASSERT(settings);
 
@@ -1948,7 +1948,7 @@ UINT32 freerdp_settings_get_uint32(const rdpSettings* settings, size_t id)
 	}
 }
 
-BOOL freerdp_settings_set_uint32(rdpSettings* settings, size_t id, UINT32 val)
+BOOL freerdp_settings_set_uint32(rdpSettings* settings, FreeRDP_Settings_Keys_UInt32 id, UINT32 val)
 {
 	union
 	{
@@ -2464,7 +2464,7 @@ BOOL freerdp_settings_set_uint32(rdpSettings* settings, size_t id, UINT32 val)
 	return TRUE;
 }
 
-INT32 freerdp_settings_get_int32(const rdpSettings* settings, size_t id)
+INT32 freerdp_settings_get_int32(const rdpSettings* settings, FreeRDP_Settings_Keys_Int32 id)
 {
 	WINPR_ASSERT(settings);
 
@@ -2484,7 +2484,7 @@ INT32 freerdp_settings_get_int32(const rdpSettings* settings, size_t id)
 	}
 }
 
-BOOL freerdp_settings_set_int32(rdpSettings* settings, size_t id, INT32 val)
+BOOL freerdp_settings_set_int32(rdpSettings* settings, FreeRDP_Settings_Keys_Int32 id, INT32 val)
 {
 	union
 	{
@@ -2516,7 +2516,7 @@ BOOL freerdp_settings_set_int32(rdpSettings* settings, size_t id, INT32 val)
 	return TRUE;
 }
 
-UINT64 freerdp_settings_get_uint64(const rdpSettings* settings, size_t id)
+UINT64 freerdp_settings_get_uint64(const rdpSettings* settings, FreeRDP_Settings_Keys_UInt64 id)
 {
 	WINPR_ASSERT(settings);
 
@@ -2533,7 +2533,7 @@ UINT64 freerdp_settings_get_uint64(const rdpSettings* settings, size_t id)
 	}
 }
 
-BOOL freerdp_settings_set_uint64(rdpSettings* settings, size_t id, UINT64 val)
+BOOL freerdp_settings_set_uint64(rdpSettings* settings, FreeRDP_Settings_Keys_UInt64 id, UINT64 val)
 {
 	union
 	{
@@ -2561,7 +2561,7 @@ BOOL freerdp_settings_set_uint64(rdpSettings* settings, size_t id, UINT64 val)
 	return TRUE;
 }
 
-INT64 freerdp_settings_get_int64(const rdpSettings* settings, size_t id)
+INT64 freerdp_settings_get_int64(const rdpSettings* settings, FreeRDP_Settings_Keys_Int64 id)
 {
 	WINPR_ASSERT(settings);
 
@@ -2575,7 +2575,7 @@ INT64 freerdp_settings_get_int64(const rdpSettings* settings, size_t id)
 	}
 }
 
-BOOL freerdp_settings_set_int64(rdpSettings* settings, size_t id, INT64 val)
+BOOL freerdp_settings_set_int64(rdpSettings* settings, FreeRDP_Settings_Keys_Int64 id, INT64 val)
 {
 	union
 	{
@@ -2599,7 +2599,8 @@ BOOL freerdp_settings_set_int64(rdpSettings* settings, size_t id, INT64 val)
 	return TRUE;
 }
 
-const char* freerdp_settings_get_string(const rdpSettings* settings, size_t id)
+const char* freerdp_settings_get_string(const rdpSettings* settings,
+                                        FreeRDP_Settings_Keys_String id)
 {
 	WINPR_ASSERT(settings);
 
@@ -2904,7 +2905,7 @@ const char* freerdp_settings_get_string(const rdpSettings* settings, size_t id)
 	}
 }
 
-char* freerdp_settings_get_string_writable(rdpSettings* settings, size_t id)
+char* freerdp_settings_get_string_writable(rdpSettings* settings, FreeRDP_Settings_Keys_String id)
 {
 	WINPR_ASSERT(settings);
 
@@ -3209,7 +3210,8 @@ char* freerdp_settings_get_string_writable(rdpSettings* settings, size_t id)
 	}
 }
 
-BOOL freerdp_settings_set_string_(rdpSettings* settings, size_t id, char* val, size_t len)
+BOOL freerdp_settings_set_string_(rdpSettings* settings, FreeRDP_Settings_Keys_String id, char* val,
+                                  size_t len)
 {
 	union
 	{
@@ -3524,12 +3526,14 @@ BOOL freerdp_settings_set_string_(rdpSettings* settings, size_t id, char* val, s
 	return TRUE;
 }
 
-BOOL freerdp_settings_set_string_len(rdpSettings* settings, size_t id, const char* val, size_t len)
+BOOL freerdp_settings_set_string_len(rdpSettings* settings, FreeRDP_Settings_Keys_String id,
+                                     const char* val, size_t len)
 {
 	return freerdp_settings_set_string_copy_(settings, id, val, len, TRUE);
 }
 
-BOOL freerdp_settings_set_string(rdpSettings* settings, size_t id, const char* val)
+BOOL freerdp_settings_set_string(rdpSettings* settings, FreeRDP_Settings_Keys_String id,
+                                 const char* val)
 {
 	size_t len = 0;
 	if (val)
@@ -3537,8 +3541,8 @@ BOOL freerdp_settings_set_string(rdpSettings* settings, size_t id, const char* v
 	return freerdp_settings_set_string_copy_(settings, id, val, len, TRUE);
 }
 
-BOOL freerdp_settings_set_string_copy_(rdpSettings* settings, size_t id, const char* val,
-                                       size_t len, BOOL cleanup)
+BOOL freerdp_settings_set_string_copy_(rdpSettings* settings, FreeRDP_Settings_Keys_String id,
+                                       const char* val, size_t len, BOOL cleanup)
 {
 	union
 	{
@@ -3858,7 +3862,7 @@ BOOL freerdp_settings_set_string_copy_(rdpSettings* settings, size_t id, const c
 	return TRUE;
 }
 
-void* freerdp_settings_get_pointer_writable(rdpSettings* settings, size_t id)
+void* freerdp_settings_get_pointer_writable(rdpSettings* settings, FreeRDP_Settings_Keys_Pointer id)
 {
 	WINPR_ASSERT(settings);
 
@@ -3965,7 +3969,8 @@ void* freerdp_settings_get_pointer_writable(rdpSettings* settings, size_t id)
 	}
 }
 
-BOOL freerdp_settings_set_pointer(rdpSettings* settings, size_t id, const void* val)
+BOOL freerdp_settings_set_pointer(rdpSettings* settings, FreeRDP_Settings_Keys_Pointer id,
+                                  const void* val)
 {
 	union
 	{
